Move buffgiverfunctionality into level folder
