+++ Game/Game.vcxproj
@@ -228,6 +228,7 @@
     <ClInclude Include="Camera\CameraFunctionality.h" />
     <ClInclude Include="Character\Buffs\CharacterBuffsFunctionality.h" />
     <ClInclude Include="Character\Stats\CharacterStatsFunctionality.h" />
+    <ClInclude Include="Levels\BuffGiverFunctionality.h" />
     <ClInclude Include="Movement\SpinnerFunctionality.h" />
     <ClInclude Include="Designers\LevelDesigner\LevelDesigner.h" />
     <ClInclude Include="Character\CharacterFunctionality.h" />
@@ -255,7 +256,6 @@
     <ClInclude Include="Player\PlayerControllerFunctionality.h" />
     <ClInclude Include="Weapons\MeleeWeapon.h" />
     <ClInclude Include="Weapons\WeaponProperty\WeaponProperty.h" />
-    <ClInclude Include="WorldObjects\BuffGiverFunctionality.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Actions\ActionFunctionality.cpp" />
@@ -264,6 +264,7 @@
     <ClCompile Include="Camera\CameraFunctionality.cpp" />
     <ClCompile Include="Character\Buffs\CharacterBuffsFunctionality.cpp" />
     <ClCompile Include="Character\Stats\CharacterStatsFunctionality.cpp" />
+    <ClCompile Include="Levels\BuffGiverFunctionality.cpp" />
     <ClCompile Include="Movement\SpinnerFunctionality.cpp" />
     <ClCompile Include="Designers\LevelDesigner\LevelDesigner.cpp" />
     <ClCompile Include="Character\CharacterFunctionality.cpp" />
@@ -293,7 +294,6 @@
     <ClCompile Include="Player\PlayerControllerFunctionality.cpp" />
     <ClCompile Include="Weapons\MeleeWeapon.cpp" />
     <ClCompile Include="Weapons\WeaponProperty\WeaponProperty.cpp" />
-    <ClCompile Include="WorldObjects\BuffGiverFunctionality.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\WonderMake\WonderMake.vcxproj">
